final integration

+++ src/Controller/ReponseController.php
@@ -24,38 +24,53 @@ class ReponseController extends AbstractController
     /**
      * @Route("/", name="all_reclamation", methods={"GET"})
      */
-    public function showBackOffice(ReclamationRepository $reclamationRepository,Request $request,PaginatorInterface $paginator): Response
+    public function showBackOffice(ReclamationRepository $reclamationRepository, Request $request, PaginatorInterface $paginator): Response
     {
-        $reclamations = $reclamationRepository->findAll();
-        $pagination = $paginator->paginate(
-            $reclamations,
-            $request->query->getInt('page', 1),
-            2
-        );
-        return $this->render('reponse/show_all.html.twig', [
-            'reclamations' => $pagination,
-        ]);
+        $user = $this->getUser();
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        } else if ($user->getRoles() == ["ROLE_ADMIN"]) {
+            $reclamations = $reclamationRepository->findAll();
+            $pagination = $paginator->paginate(
+                $reclamations,
+                $request->query->getInt('page', 1),
+                2
+            );
+            return $this->render('reponse/show_all.html.twig', [
+                'reclamations' => $pagination,
+            ]);
+        } else {
+            return $this->redirectToRoute('error');
+        }
+
     }
 
     /**
      * @Route("/noreply", name="no_reply_reclamation", methods={"GET", "POST"})
      * @throws Exception
      */
-    public function notifications(ReclamationRepository $reclamationRepository, Request $req,PaginatorInterface $paginator): Response
+    public function notifications(ReclamationRepository $reclamationRepository, Request $req, PaginatorInterface $paginator): Response
     {
-        $data = $reclamationRepository->findBy(array("status"=>"0"));
-        $pagination = $paginator->paginate(
-            $data,
-            $req->query->getInt('page',1),
-            2
-        );
-        $nbNull = $reclamationRepository->getNonTreatedReports();
-        $nb = $reclamationRepository->getTreatedReports();
-        return $this->render('reclamation/noreply.html.twig',[
-            'noreply'=>$pagination,
-            'nbNull'=>$nbNull,
-            'nb'=>$nb
-        ]);
+        $user = $this->getUser();
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        } else if ($user->getRoles() == ["ROLE_ADMIN"]) {
+            $data = $reclamationRepository->findBy(array("status" => "0"));
+            $pagination = $paginator->paginate(
+                $data,
+                $req->query->getInt('page', 1),
+                2
+            );
+            $nbNull = $reclamationRepository->getNonTreatedReports();
+            $nb = $reclamationRepository->getTreatedReports();
+            return $this->render('reclamation/noreply.html.twig', [
+                'noreply' => $pagination,
+                'nbNull' => $nbNull,
+                'nb' => $nb
+            ]);
+        } else {
+            return $this->redirectToRoute('error');
+        }
     }
 
     /**
@@ -63,22 +78,29 @@ class ReponseController extends AbstractController
      */
     public function editReponse(Request $request, Reponse $reponse, Reclamation $rec, EntityManagerInterface $entityManager): Response
     {
-        $form = $this->createForm(ReponseType::class, $reponse);
-        $formRec = $this->createForm(ReclamationType::class, $rec);
-        $form->handleRequest($request);
+        $user = $this->getUser();
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        } else if ($user->getRoles() == ["ROLE_ADMIN"]) {
+            $form = $this->createForm(ReponseType::class, $reponse);
+            $formRec = $this->createForm(ReclamationType::class, $rec);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $reponse->setReplyDate(new \DateTime());
-            $entityManager->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $reponse->setReplyDate(new \DateTime());
+                $entityManager->flush();
 
-            return $this->redirectToRoute('all_reclamation', [], Response::HTTP_SEE_OTHER);
+                return $this->redirectToRoute('all_reclamation', [], Response::HTTP_SEE_OTHER);
+            }
+
+            return $this->render('reponse/new.html.twig', [
+                'reponse' => $reponse,
+                'form' => $form->createView(),
+                'form_rec' => $formRec->createView(),
+            ]);
+        } else {
+            return $this->redirectToRoute('error');
         }
-
-        return $this->render('reponse/new.html.twig', [
-            'reponse' => $reponse,
-            'form' => $form->createView(),
-            'form_rec' => $formRec->createView(),
-        ]);
     }
 
     /**
@@ -86,59 +108,67 @@ class ReponseController extends AbstractController
      */
     public function add(Request $request, Reclamation $rec, EntityManagerInterface $entityManager, \Swift_Mailer $mailer): Response
     {
-        $reponse = new Reponse();
-        $formRec = $this->createForm(ReclamationType::class, $rec);
-        $form = $this->createForm(ReponseType::class, $reponse);
-        $form->handleRequest($request);
+        $user = $this->getUser();
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        } else if ($user->getRoles() == ["ROLE_ADMIN"]) {
+            $reponse = new Reponse();
+            $formRec = $this->createForm(ReclamationType::class, $rec);
+            $form = $this->createForm(ReponseType::class, $reponse);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $file = $form->get('admin_file')->getData();
-            $email = $formRec->get('email')->getData();
-            $name = $formRec->get('name')->getData();
-            $reply = $form->get('answer')->getData();
-            if ($file) {
-                // this is needed to safely include the file name as part of the URL
-                $newFilename = uniqid().'.'.$file->guessExtension();
-                // Move the file to the directory where pictures are stored
-                try {
-                    $file->move(
-                        $this->getParameter('reclamation_file_dir'),
-                        $newFilename
-                    );
-                } catch (FileException $e) {
-                    // ... handle exception if something happens during file upload
+            if ($form->isSubmitted() && $form->isValid()) {
+                $file = $form->get('admin_file')->getData();
+                $email = $formRec->get('email')->getData();
+                $name = $formRec->get('name')->getData();
+                $reply = $form->get('answer')->getData();
+                if ($file) {
+                    // this is needed to safely include the file name as part of the URL
+                    $newFilename = uniqid() . '.' . $file->guessExtension();
+                    // Move the file to the directory where pictures are stored
+                    try {
+                        $file->move(
+                            $this->getParameter('reclamation_file_dir'),
+                            $newFilename
+                        );
+                    } catch (FileException $e) {
+                        // ... handle exception if something happens during file upload
+                    }
+
+                    // updates the 'brochureFilename' property to store the PDF file name
+                    // instead of its contents
+                    $reponse->setAdminFile($newFilename);
+                    $message = (new \Swift_Message('Réclamation'))
+                        ->setFrom('alemnicontact@gmail.com')
+                        ->setTo($email)
+                        ->setBody(
+                            $this->render('emails/reclamation.html.twig', [
+                                'name' => $name,
+                                'reply' => $reply
+                            ]),
+                            'text/html'
+                        );
+                    $mailer->send($message);
                 }
+                $reponse->setReclamation($rec);
+                $reponse->setReplyDate(new \DateTime());
+                $rec->setStatus('answered');
+                $entityManager->persist($rec);
+                $entityManager->persist($reponse);
+                $entityManager->flush();
 
-                // updates the 'brochureFilename' property to store the PDF file name
-                // instead of its contents
-                $reponse->setAdminFile($newFilename);
-                $message = (new \Swift_Message('Réclamation'))
-                    ->setFrom('alemnicontact@gmail.com')
-                    ->setTo($email)
-                    ->setBody(
-                        $this->render('emails/reclamation.html.twig',[
-                            'name'=>$name,
-                            'reply'=>$reply
-                        ]),
-                        'text/html'
-                    );
-                $mailer->send($message);
+                return $this->redirectToRoute('all_reclamation', [], Response::HTTP_SEE_OTHER);
             }
-            $reponse->setReclamation($rec);
-            $reponse->setReplyDate(new \DateTime());
-            $rec->setStatus('answered');
-            $entityManager->persist($rec);
-            $entityManager->persist($reponse);
-            $entityManager->flush();
 
-            return $this->redirectToRoute('all_reclamation', [], Response::HTTP_SEE_OTHER);
+            return $this->render('reponse/new.html.twig', [
+                'reponse' => $reponse,
+                'form' => $form->createView(),
+                'form_rec' => $formRec->createView(),
+            ]);
+        } else {
+            return $this->redirectToRoute('error');
         }
 
-        return $this->render('reponse/new.html.twig', [
-            'reponse' => $reponse,
-            'form' => $form->createView(),
-            'form_rec' => $formRec->createView(),
-        ]);
     }
 
     /**
@@ -146,11 +176,19 @@ class ReponseController extends AbstractController
      */
     public function delete(Request $request, Reponse $reponse, EntityManagerInterface $entityManager): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$reponse->getId(), $request->request->get('_token'))) {
-            $entityManager->remove($reponse);
-            $entityManager->flush();
+        $user = $this->getUser();
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        } else if ($user->getRoles() == ["ROLE_ADMIN"]) {
+            if ($this->isCsrfTokenValid('delete' . $reponse->getId(), $request->request->get('_token'))) {
+                $entityManager->remove($reponse);
+                $entityManager->flush();
+            }
+
+            return $this->redirectToRoute('all_reclamation', [], Response::HTTP_SEE_OTHER);
+        } else {
+            return $this->redirectToRoute('error');
         }
 
-        return $this->redirectToRoute('all_reclamation', [], Response::HTTP_SEE_OTHER);
     }
 }
