+++ src/Controller/SeanceController.php
@@ -3,9 +3,11 @@
 namespace App\Controller;
 
 use App\Entity\Formation;
+use App\Entity\Quiz;
 use App\Entity\Seance;
 use App\Form\SeanceType;
 use App\Repository\FormationRepository;
+use App\Repository\QuizRepository;
 use App\Repository\SeanceRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -31,69 +33,92 @@ class SeanceController extends AbstractController
     /**
      * @Route("/{id}/editSeance", name="seance1_edit", methods={"GET", "POST"})
      */
-    public function editSeance(Request $request , Seance $seance, EntityManagerInterface $entityManager,SeanceRepository $seanceRepository): Response
+    public function editSeance(Request $request, Seance $seance, EntityManagerInterface $entityManager, SeanceRepository $seanceRepository): Response
     {
-        $form = $this->createForm(SeanceType::class, $seance);
-        $form->handleRequest($request);
+        $user = $this->getUser();
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        } else if ($user->getRoles() == ["ROLE_TUTOR"]) {
+            $form = $this->createForm(SeanceType::class, $seance);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $entityManager->flush();
 
-            return $this->redirectToRoute('formation_new', [], Response::HTTP_SEE_OTHER);
+                return $this->redirectToRoute('formation_new', [], Response::HTTP_SEE_OTHER);
+            }
+
+            return $this->render('seance/edit.html.twig', [
+                'seance' => $seance,
+                'form' => $form->createView(),
+                'seances' => $seanceRepository->findByExampleField($seance->getFormation()->getId()),
+            ]);
+        } else {
+            return $this->redirectToRoute('error');
         }
 
-        return $this->render('seance/edit.html.twig', [
-            'seance' => $seance,
-            'form' => $form->createView(),
-            'seances' => $seanceRepository->findByExampleField($seance->getFormation()->getId()),
-        ]);
     }
+
     /**
      * @Route("/new/{id}", name="seance_new", methods={"GET", "POST"})
      */
-    public function new(Request $request,Formation $formation, EntityManagerInterface $entityManager,FormationRepository $formationRepository,SeanceRepository $seanceRepository): Response
+    public function new(Request $request, Formation $formation, EntityManagerInterface $entityManager, FormationRepository $formationRepository, SeanceRepository $seanceRepository): Response
     {
-        $seance = new Seance();
-        $form = $this->createForm(SeanceType::class, $seance);
-        $form->handleRequest($request);
+        $user = $this->getUser();
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        } else if ($user->getRoles() == ["ROLE_TUTOR"]) {
+            $seance = new Seance();
+            $form = $this->createForm(SeanceType::class, $seance);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-           #$test= $this->getDoctrine()->getRepository(Formation::class)->find('delete'.$formation.);
-            $seance->setFormation($formation);
+            if ($form->isSubmitted() && $form->isValid()) {
+                #$test= $this->getDoctrine()->getRepository(Formation::class)->find('delete'.$formation.);
+                $seance->setFormation($formation);
 
-            $entityManager->persist($seance);
-            $entityManager->flush();
+                $entityManager->persist($seance);
+                $entityManager->flush();
 
-            return $this->redirectToRoute('formation_new', [], Response::HTTP_SEE_OTHER);
-           }
+                return $this->redirectToRoute('formation_new', [], Response::HTTP_SEE_OTHER);
+            }
 
 
-        return $this->render('seance/new.html.twig', [
-            'seance' => $seance,
-            'form' => $form->createView(),
-            'seances' => $seanceRepository->findByExampleField($formation->getId()),
-        ]);
+            return $this->render('seance/new.html.twig', [
+                'seance' => $seance,
+                'form' => $form->createView(),
+                'seances' => $seanceRepository->findByExampleField($formation->getId()),
+            ]);
+        } else {
+            return $this->redirectToRoute('error');
+        }
     }
 
     /**
      * @Route("/{id}/edit", name="seance_edit", methods={"GET", "POST"}, requirements={"id":"\d+"})
      */
-    public function edit(Request $request, Formation $formation, Seance $seance, EntityManagerInterface $entityManager,SeanceRepository $seanceRepository): Response
+    public function edit(Request $request, Formation $formation, Seance $seance, EntityManagerInterface $entityManager, SeanceRepository $seanceRepository): Response
     {
-        $form = $this->createForm(SeanceType::class, $seance);
-        $form->handleRequest($request);
+        $user = $this->getUser();
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        } else if ($user->getRoles() == ["ROLE_TUTOR"]) {
+            $form = $this->createForm(SeanceType::class, $seance);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $entityManager->flush();
 
-            return $this->redirectToRoute('formation_new', [], Response::HTTP_SEE_OTHER);
+                return $this->redirectToRoute('formation_new', [], Response::HTTP_SEE_OTHER);
+            }
+
+            return $this->render('seance/edit.html.twig', [
+                'seance' => $seance,
+                'form' => $form->createView(),
+                'seances' => $seanceRepository->findByExampleField($formation->getId()),
+            ]);
+        } else {
+            return $this->redirectToRoute('error');
         }
-
-        return $this->render('seance/edit.html.twig', [
-            'seance' => $seance,
-            'form' => $form->createView(),
-            'seances' => $seanceRepository->findByExampleField($formation->getId()),
-        ]);
     }
 
     /**
@@ -109,13 +134,29 @@ class SeanceController extends AbstractController
     /**
      * @Route("/{id}", name="seance_delete", methods={"POST"})
      */
-    public function delete(Request $request, Seance $seance, EntityManagerInterface $entityManager): Response
+    public function delete(Request $request, Seance $seance, EntityManagerInterface $entityManager, QuizRepository $quizRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$seance->getId(), $request->request->get('_token'))) {
-            $entityManager->remove($seance);
-            $entityManager->flush();
+        $user = $this->getUser();
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        } else if ($user->getRoles() == ["ROLE_TUTOR"]) {
+            if ($this->isCsrfTokenValid('delete' . $seance->getId(), $request->request->get('_token'))) {
+                $entityManager->remove($seance);
+                $entityManager->flush();
+            }
+            if ($seance->getFormation()->getQuiz() != null) {
+                $quiz = $quizRepository->find($seance->getFormation()->getQuiz()->getId());
+                if ($seance->getFormation()->getSeances()->count() == 0) {
+                    if ($seance->getFormation()->getQuiz() != null) {
+                        $entityManager->remove($quiz);
+                        $entityManager->flush();
+                    }
+                }
+            }
+            return $this->redirectToRoute('formation_new', [], Response::HTTP_SEE_OTHER);
+        } else {
+            return $this->redirectToRoute('error');
         }
 
-        return $this->redirectToRoute('formation_new', [], Response::HTTP_SEE_OTHER);
     }
 }
